+++ src/features/mainpage/index.ts
@@ -4,5 +4,5 @@ export * from "./ScheduleSection";
 export * from "./LetterSection";
 export * from "./DDaySection";
 export * from "./SpecialDateSection";
-export * from "./utils/EmotionSelector";
-export * from "./utils/MainHeader";
+export * from "./ui/EmotionSelector";
+export * from "./ui/MainHeader";
