+++ src/features/mainpage/model/InitSettingStore.ts
