+++ src/features/mainpage/ui/EmotionSelector.tsx
