+++ src/features/mypage/index.ts
@@ -1,6 +1,6 @@
-export * from "./AlarmSection";
-export * from "./ProfileCard";
+export * from "./AlarmSettingSection";
+export * from "./ProfileCardSection";
 export * from "./StatusSection";
-export * from "./widgets";
+export * from "./ui";
 export * from "./EditInfoCard";
 export * from "./AlarmToggleList";
