--- src/pages/couplematch/index.ts
@@ -1,1 +1,0 @@
-export * from "./CoupleContact";
