+++ src/pages/mainPage/initpage/index.ts
@@ -1,0 +1,3 @@
+export * from "./FirstMeet";
+export * from "./MilitaryDay";
+export * from "./Where";
