--- src/pages/onboarding/Onboarding.tsx
@@ -1,3 +1,0 @@
-export const Onboarding = () => {
-  return <div>onboarding</div>;
-};
